Fix package reference to system dacpac
Package references expect the dacpac name to be the same as package name, but is not the case for system dacpac packages. This change adds condition to check for `Microsoft.SqlServer.Dacpacs` package references and redirect to the master.dacpac within. Also adds `DacpacName` as an overridable atttribute to the PackageReference.